refactor(crypto): validate max random bytes in Rust (#10857)
