+++ extensions/crypto/Cargo.toml
@@ -15,5 +15,6 @@ path = "lib.rs"
 
 [dependencies]
 deno_core = { version = "0.88.1", path = "../../core" }
+deno_web = { version = "0.38.1", path = "../web" }
 rand = "0.8.3"
 uuid = { version = "0.8.2", features = ["v4"] }
